♻️ Respect user prefs for timestamps
