+++ static/app/components/events/breadcrumbs/breadcrumbsTimeline.tsx
@@ -1,15 +1,19 @@
 import {Fragment, useRef} from 'react';
 import styled from '@emotion/styled';
 import {useVirtualizer} from '@tanstack/react-virtual';
+import moment from 'moment';
 
+import DateTime from 'sentry/components/dateTime';
+import Duration from 'sentry/components/duration';
 import BreadcrumbsItemContent from 'sentry/components/events/breadcrumbs/breadcrumbsItemContent';
-import {
-  BREADCRUMB_TIMESTAMP_PLACEHOLDER,
-  type EnhancedCrumb,
-} from 'sentry/components/events/breadcrumbs/utils';
-import Timeline, {type TimelineItemProps} from 'sentry/components/timeline';
+import type {EnhancedCrumb} from 'sentry/components/events/breadcrumbs/utils';
+import Timeline from 'sentry/components/timeline';
+import {Tooltip} from 'sentry/components/tooltip';
 import {t} from 'sentry/locale';
+import {space} from 'sentry/styles/space';
 import {defined} from 'sentry/utils';
+import isValidDate from 'sentry/utils/date/isValidDate';
+import {shouldUse24Hours} from 'sentry/utils/dates';
 
 interface BreadcrumbsTimelineProps {
   breadcrumbs: EnhancedCrumb[];
@@ -22,7 +26,10 @@ interface BreadcrumbsTimelineProps {
    * Useful for connecting timeline to components rendered after it.
    */
   showLastLine?: boolean;
-  startTimeString?: TimelineItemProps['startTimeString'];
+  /**
+   * If specified, will display time relatively.
+   */
+  startTimeString?: string;
 }
 
 export default function BreadcrumbsTimeline({
@@ -50,6 +57,29 @@ export default function BreadcrumbsTimeline({
     const {breadcrumb, raw, title, meta, iconComponent, colorConfig, levelComponent} =
       breadcrumbs[virtualizedRow.index];
     const isVirtualCrumb = !defined(raw);
+
+    const timeDate = new Date(breadcrumb.timestamp ?? '');
+    const startTimeDate = new Date(startTimeString ?? '');
+
+    const absoluteFormat = shouldUse24Hours() ? 'HH:mm:ss.SSS' : 'hh:mm:ss.SSS';
+    const timestampComponent = isValidDate(timeDate) ? (
+      <Timestamp>
+        <Tooltip
+          title={<DateTime date={timeDate} format={`ll - ${absoluteFormat} (z)`} />}
+        >
+          {isValidDate(startTimeDate) ? (
+            <Duration
+              seconds={moment(timeDate).diff(moment(startTimeDate), 's', true)}
+              exact
+              abbreviation
+            />
+          ) : (
+            <DateTime date={timeDate} format={absoluteFormat} />
+          )}
+        </Tooltip>
+      </Timestamp>
+    ) : null;
+
     return (
       <Timeline.Item
         key={virtualizedRow.key}
@@ -63,8 +93,7 @@ export default function BreadcrumbsTimeline({
         }
         colorConfig={colorConfig}
         icon={iconComponent}
-        timeString={breadcrumb.timestamp ?? BREADCRUMB_TIMESTAMP_PLACEHOLDER}
-        startTimeString={startTimeString}
+        timestamp={timestampComponent}
         // XXX: Only the virtual crumb can be marked as active for breadcrumbs
         isActive={isVirtualCrumb ?? false}
         style={showLastLine ? {background: 'transparent'} : {}}
@@ -84,7 +113,6 @@ export default function BreadcrumbsTimeline({
       ref={containerRef}
       style={{
         height: virtualizer.getTotalSize(),
-
         contain: 'layout size',
       }}
     >
@@ -98,4 +126,13 @@ const Subtitle = styled('p')`
   font-weight: normal;
   font-size: ${p => p.theme.fontSizeSmall};
   display: inline;
+`;
+
+const Timestamp = styled('div')`
+  margin: 0 ${space(1)};
+  color: ${p => p.theme.subText};
+  font-size: ${p => p.theme.fontSizeSmall};
+  span {
+    text-decoration: underline dashed ${p => p.theme.subText};
+  }
 `;
